(feat) allow adding non mpcFill images as backs
Allow any image placed in the images folder to be added as a card, both front and back. They can be added as the
back to a card from the xml or to another card added via the images folder. This is achieved by prompting the
user on each card found that isn't already assigned in the xml, then adding it to the xml.